Modifying..

+++ src/main/java/com/IronLibrary/HW3IronLibrary/repository/AuthorRepository.java
@@ -1,7 +1,7 @@
 package com.IronLibrary.HW3IronLibrary.repository;
 
-import com.IronLibrary.HW3IronLibrary.Entity.Author;
+import com.IronLibrary.HW3IronLibrary.entity.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepository extends JpaRepository<Author, Long> {
+public interface AuthorRepository extends JpaRepository<Author, Integer> {
 }
